seperate schema loading and resource loading logic from validation
