--- data/resources/notifications-integrations/config.yaml
@@ -1,3 +1,0 @@
-resource_type: notifications/integration
-resource_reporters:
-  - NOTIFICATIONS
