+++ internal/middleware/validation.go
@@ -6,19 +6,14 @@ import (
 	"fmt"
 	"github.com/bufbuild/protovalidate-go"
 	"github.com/go-kratos/kratos/v2/errors"
-	"github.com/go-kratos/kratos/v2/log"
 	"github.com/go-kratos/kratos/v2/middleware"
 	"google.golang.org/protobuf/encoding/protojson"
 	"google.golang.org/protobuf/proto"
-	"gopkg.in/yaml.v3"
 	"os"
-	"path/filepath"
 	"strings"
 
 	"github.com/xeipuuv/gojsonschema"
 )
-
-const defaultResourceDir = "data/resources"
 
 var (
 	resourceDir          = os.Getenv("RESOURCE_DIR")
@@ -27,22 +22,7 @@ var (
 )
 
 func Validation(validator *protovalidate.Validator) middleware.Middleware {
-	if resourceDir == "" {
-		resourceDir = defaultResourceDir
-	}
-	resourceDirs, err := os.ReadDir(resourceDir)
-	if err != nil {
-		log.Fatalf("Failed to read resource directory %s: %v", resourceDir, err)
-	}
-	log.Infof("Read resource directory %s:", resourceDir)
-
-	for _, dir := range resourceDirs {
-		if !dir.IsDir() {
-			continue
-		}
-
-		AllowedResourceTypes[dir.Name()] = struct{}{}
-	}
+	LoadResources()
 	return func(handler middleware.Handler) middleware.Handler {
 		return func(ctx context.Context, req interface{}) (interface{}, error) {
 			if v, ok := req.(proto.Message); ok {
@@ -50,7 +30,7 @@ func Validation(validator *protovalidate.Validator) middleware.Middleware {
 				if err := validator.Validate(v); err != nil {
 					return nil, errors.BadRequest("VALIDATOR", err.Error()).WithCause(err)
 				}
-				if err := ValidateResourceJSON(v); err != nil {
+				if err := validateResourceJSON(v); err != nil {
 					return nil, errors.BadRequest("JSON_VALIDATOR", err.Error()).WithCause(err)
 				}
 			}
@@ -59,7 +39,7 @@ func Validation(validator *protovalidate.Validator) middleware.Middleware {
 	}
 }
 
-func ValidateResourceJSON(msg proto.Message) error {
+func validateResourceJSON(msg proto.Message) error {
 	data, err := protojson.Marshal(msg)
 	if err != nil {
 		return fmt.Errorf("failed to marshal message: %w", err)
@@ -99,7 +79,7 @@ func ValidateResourceJSON(msg proto.Message) error {
 		return fmt.Errorf("resource_type '%s' is abstract and cannot have resource_data", resourceTypeMetadata)
 	} else {
 		// Validate resource_data if not abstract
-		resourceSchema, err := loadSchema(resourceTypeMetadata)
+		resourceSchema, err := LoadSchema(resourceTypeMetadata)
 		if err != nil {
 			return fmt.Errorf("failed to load schema for '%s': %w", resourceTypeMetadata, err)
 		}
@@ -123,7 +103,7 @@ func ValidateResourceJSON(msg proto.Message) error {
 		return fmt.Errorf("resource-reporter compatibility validation failed for resource '%s': %w", resourceType, err)
 	}
 
-	reporterSchema, err := loadReporterSchema(resourceTypeMetadata, strings.ToLower(reporterType))
+	reporterSchema, err := LoadReporterSchema(resourceTypeMetadata, strings.ToLower(reporterType))
 	if err != nil {
 		return fmt.Errorf("failed to load reporter schema for '%s': %w", reporterType, err)
 	}
@@ -152,58 +132,4 @@ func validateJSONAgainstSchema(schemaStr string, jsonData interface{}) error {
 		return fmt.Errorf("validation failed: %s", strings.Join(errMsgs, "; "))
 	}
 	return nil
-}
-
-func ValidateCombination(resourceType, reporterType string) error {
-	resourceReporters, err := loadValidReporters(resourceType)
-	if err != nil {
-		return fmt.Errorf("failed to load valid reporters for '%s': %w", resourceType, err)
-	}
-
-	// check if the resources reporter_type exists in the list of resource_reporters
-	for _, validReporter := range resourceReporters {
-		if reporterType == validReporter {
-			return nil
-		}
-	}
-	return fmt.Errorf("invalid reporter_type: %s for resource_type: %s", reporterType, resourceType)
-}
-
-func loadSchema(resourceType string) (string, error) {
-	schemaPath := filepath.Join(resourceDir, resourceType, resourceType+".json")
-	data, err := os.ReadFile(schemaPath)
-	if err != nil {
-		return "", fmt.Errorf("failed to read schema file for '%s': %w", resourceType, err)
-	}
-	return string(data), nil
-}
-
-func loadReporterSchema(resourceType string, reporterType string) (string, error) {
-	schemaPath := filepath.Join(resourceDir, resourceType, "reporters", reporterType, reporterType+".json")
-	data, err := os.ReadFile(schemaPath)
-	if err != nil {
-		return "", fmt.Errorf("failed to read schema file for '%s' and reporter '%s': %w", resourceType, reporterType, err)
-	}
-	return string(data), nil
-}
-
-func loadValidReporters(resourceType string) ([]string, error) {
-	configPath := filepath.Join(resourceDir, resourceType, "config.yaml")
-	data, err := os.ReadFile(configPath)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read config file for '%s': %w", resourceType, err)
-	}
-
-	var config struct {
-		ResourceReporters []string `yaml:"resource_reporters"`
-	}
-	if err := yaml.Unmarshal(data, &config); err != nil {
-		return nil, fmt.Errorf("failed to unmarshal config for '%s': %w", resourceType, err)
-	}
-
-	if config.ResourceReporters == nil {
-		return nil, fmt.Errorf("missing 'resource_reporters' field in config for '%s'", resourceType)
-	}
-
-	return config.ResourceReporters, nil
 }
